She adjustments for parallel running
Will have to adjust these more later. Maybe check for available memory
first and then decide on how many workers to run? Maybe we can switch
back to int16 again, not sure if it makes a difference though…

+++ code/Utils/subtractMean.m
@@ -2,8 +2,8 @@ function [ data ] = subtractMean( data )
 %SUBTRACTMEAN Subtracts the mean-frame of 3D-array data (width x height x 
 %frames) from each frame of data.
 
-data = int16(data);
-dataMean = int16(mean(data,3));
+data = int8(data);
+dataMean = int8(mean(data,3));
 data = data - repmat(dataMean,[1,1,size(data,3)]);
 
 end
